requirements.txt
PyQt6>=6.4.0
aiosqlite>=0.19.0
pandas>=2.0.0
aiohttp>=3.8.0
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
openpyxl>=3.1.0  # For Excel support 